Python3/Tornado/apps/pg/PG_Client/requirements.txt
asgiref==3.2.9
attrs==19.3.0
base58==2.0.1
bitarray==1.2.2
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.10.1
Django==3.0.5
django-cors-headers==3.4.0
django-filter==2.3.0
django-redis==4.12.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
drfdocs==0.0.11
eth-abi==2.1.1
eth-account==0.5.2
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.9.0
hexbytes==0.2.1
idna==2.9
importlib-metadata==1.7.0
ipfshttpclient==0.4.13.2
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
lru-dict==1.1.6
MarkupSafe==1.1.1
multiaddr==0.0.9
mysqlclient==1.4.6
netaddr==0.7.20
parsimonious==0.8.1
protobuf==3.12.2
pycryptodome==3.9.8
PyJWT==1.7.1
pyrsistent==0.16.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
rlp==1.2.0
six==1.15.0
sqlparse==0.3.1
toolz==0.10.0
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
varint==1.0.2
web3==5.9.0
websockets==8.1
zipp==3.1.0
